gh-143 Move change paper tests
